requirements.txt
llvmlite==0.38.1
numba==0.55.2
numpy==1.22.4
Pillow==9.2.0
progress==1.6
